+++ src/main/scala/hubmodel/mvmtmodels/NOMAD/NOMADModel.scala
@@ -164,8 +164,8 @@ class NOMADModel(sim: SFGraphSimulator) extends Action {
       p.travelTime = sim.currentTime - p.entryTime
 
       // sets the increments in position and velocity for the pedestrian
-      p.velocityIncrement = totalAcceleration * sim.sf_dt.value
-      p.positionIncrement = p.currentVelocity * sim.sf_dt.value
+      p.velocityIncrement = totalAcceleration * sim.sf_dt.value.toDouble
+      p.positionIncrement = p.currentVelocity * sim.sf_dt.value.toDouble
 
       //println(p.velocityIncrement , p.positionIncrement)
     }
